playground: fix compilation on macos

+++ playground/sys/src/ref/cache_block.hpp
@@ -37,10 +37,10 @@ struct cache_block_t {
   virtual void set_byte_mask(mem_access_byte_mask_t byte_mask) = 0;
   virtual mem_access_byte_mask_t get_dirty_byte_mask() const = 0;
   virtual mem_access_sector_mask_t get_dirty_sector_mask() const = 0;
-  virtual unsigned long get_last_access_time() const = 0;
+  virtual uint64_t get_last_access_time() const = 0;
   virtual void set_last_access_time(unsigned long long time,
                                     mem_access_sector_mask_t sector_mask) = 0;
-  virtual unsigned long long get_alloc_time() const = 0;
+  virtual uint64_t get_alloc_time() const = 0;
   virtual void set_ignore_on_fill(bool m_ignore,
                                   mem_access_sector_mask_t sector_mask) = 0;
   virtual void set_modified_on_fill(bool m_modified,
